Versao funcional do Wizard que precisa de ajustes
