+++ pages/student/take_course/finish_take.vue
@@ -13,7 +13,7 @@
 </template>
 <router>
     {
-        path : '/finish'
+        path : '/aluno/curso/:courseSlug/fim'
     }
 </router>
 <script>
